Move update template file access helpers
